Implementation of the depth of discharge for the storage systems in all optimization problems and rule-based operation. Minor fix to fix the dtype of numpy arrays in OpSchedule

+++ src/shipp/timeseries.py
@@ -18,7 +18,7 @@ class TimeSeries:
         if data is None:
             self.data = None
         else:
-            self.data = np.array(data)
+            self.data = np.array(data,  dtype=float)
 
         self.dt = dt
